Added Player Flower and Saw Art
I added the player flower art, including the sprites for each level of health. I  made it have the animated boil line effect using a shader. I edited the player code so that it will change sprites every -10 levels of health. I also added a script to the sprite so it spins.
The saw weapon isn't being used right now so I just put the sprite in the assets folder.