Added turrets, standardized boss/turret behavior
Movement is now also done with a finite state machine. This is so that patterns can be made for a generic "boss" and reused whenever we want. Just put down the starting attack and movement numbers and it will use the one you want. Turrets have been added to the main level as well, though they are significantly too fast at the moment. Their patterns need to be toned down to compensate.