Small Enemy that Orbits player and shoots
The code for the shooting might be a little bit rough, but I will make certain it works over the weekend. Right now it's 11:25 and I am out of Monster and kinda want to be done a little bit. But it definitely orbits the player pretty closely, will add some noise to that at some point.